hw10_program_optimization/stats_test.go
//go:build !bench
// +build !bench

package hw10programoptimization

import (
	"bytes"
	"testing"

	"github.com/stretchr/testify/require"
)

func TestGetDomainStat(t *testing.T) {
	data := `{"Id":1,"Name":"Howard Mendoza","Username":"0Oliver","Email":"aliquid_qui_ea@Browsedrive.gov","Phone":"6-866-899-36-79","Password":"InAQJvsq","Address":"Blackbird Place 25"}
{"Id":2,"Name":"Jesse Vasquez","Username":"qRichardson","Email":"mLynch@broWsecat.com","Phone":"9-373-949-64-00","Password":"SiZLeNSGn","Address":"Fulton Hill 80"}
{"Id":3,"Name":"Clarence Olson","Username":"RachelAdams","Email":"RoseSmith@Browsecat.com","Phone":"988-48-97","Password":"71kuz3gA5w","Address":"Monterey Park 39"}
{"Id":4,"Name":"Gregory Reid","Username":"tButler","Email":"5Moore@Teklist.net","Phone":"520-04-16","Password":"r639qLNu","Address":"Sunfield Park 20"}
{"Id":5,"Name":"Janice Rose","Username":"KeithHart","Email":"nulla@Linktype.com","Phone":"146-91-01","Password":"acSBF5","Address":"Russell Trail 61"}`

	t.Run("find 'com'", func(t *testing.T) {
		result, err := GetDomainStat(bytes.NewBufferString(data), "com")
		require.NoError(t, err)
		require.Equal(t, DomainStat{
			"browsecat.com": 2,
			"linktype.com":  1,
		}, result)
	})

	t.Run("find 'gov'", func(t *testing.T) {
		result, err := GetDomainStat(bytes.NewBufferString(data), "gov")
		require.NoError(t, err)
		require.Equal(t, DomainStat{"browsedrive.gov": 1}, result)
	})

	t.Run("find 'unknown'", func(t *testing.T) {
		result, err := GetDomainStat(bytes.NewBufferString(data), "unknown")
		require.NoError(t, err)
		require.Equal(t, DomainStat{}, result)
	})
}

func Test_appendStat(t *testing.T) {
	type args struct {
		domain string
		bytes  []byte
		stat   *DomainStat
	}
	tests := []struct {
		name     string
		args     args
		wantErr  bool
		wantStat DomainStat
	}{
		{
			name: "correct add stat",
			args: args{
				domain: "ru",
				bytes:  []byte(`{"Id":1,"Name":"Howard Mendoza","Username":"0Oliver","Email":"aliquid_qui_ea@mail.ru","Phone":"6-866-899-36-79","Password":"InAQJvsq","Address":"Blackbird Place 25"}`),
				stat: &DomainStat{
					"mail.ru":    2,
					"google.com": 2,
				},
			},
			wantErr: false,
			wantStat: DomainStat{
				"mail.ru":    3,
				"google.com": 2,
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := appendStat(tt.args.domain, tt.args.bytes, tt.args.stat)
			if (err != nil) != tt.wantErr {
				t.Errorf("appendStat() error = %v, wantErr %v", err, tt.wantErr)
			}
			for k, v := range tt.wantStat {
				require.Equal(t, (*tt.args.stat)[k], v)
			}
		})
	}
}
